Move CCLF import functions into new cli.go; expand tests
